Working greedy optimization with one quirk tweak that may or may not prove efficient for a heavier optimization: Pmax and Pmin are randomized for each use to allow for broad and narrow ranges. This proved to be faster for learning a simple function. I.e. fitting data points to a function.

+++ proto/greedy-optimization/Assets/UniformDistribution.cs
@@ -9,9 +9,9 @@ public class UniformDistribution
         p_fixedrand = new System.Random(4350809);
     }
 
-    public float U(double p_lower, double p_upper)
+    public double U(double p_lower, double p_upper)
     {
-        float u = (float)(p_fixedrand.NextDouble() * (p_upper - p_lower) + p_lower);
+        double u = p_fixedrand.NextDouble() * (p_upper - p_lower) + p_lower;
         return u;
     }
 }
